SHOW IMAGES
We downloaded some images and put them in a JLabel, then we saw how to import an image to a package as a JLabel icon. We had the problem that the image was too large and enlarged to fit inside the JLabel, so we had to resize the image to fit the JLabel. We put a jpeg image in the spdvi folder and in the property tab of the JLabel we included that image. We added another JLabel and selected another image that was not in our package, the program showed it to us but there is a problem, and that is that it is using the absolute path of "C: \ Users \ David \ .... \ image1.jpeg" , and if you want to see that app from other computers, it will not load that image. As I mentioned before, the images were not shown complete because their size (in pixels) was larger than that of the JLabel. Through a method we change the size of the images without changing their aspect ratio. We use classes like: BufferedImage and ImageIO. We use a BufferedImage to read the image. We resize the image using DEFAULT and SMOOTH. Then we calculate the width and height and clicking the button we reconvert the image.